Refactor backend architecture and add project creation loading overlay
