+++ backend/app/routes/__init__.py
@@ -1,0 +1,4 @@
+from .auth import router as auth_router
+from .drive import router as drive_router
+
+__all__ = ["auth_router", "drive_router"]
