+++ backend/app/routes/drive.py
@@ -1,0 +1,44 @@
+from __future__ import annotations
+
+from typing import Any, Dict, List, Optional
+
+from fastapi import APIRouter, File, Form, HTTPException, Query, UploadFile
+from fastapi.responses import JSONResponse
+
+from ..container import drive_service
+
+router = APIRouter()
+
+
+@router.post("/drive/gs/setup")
+async def ensure_gs_folder(
+    google_id: Optional[str] = Query(None, description="Drive 작업에 사용할 Google 사용자 식별자 (sub)"),
+) -> JSONResponse:
+    result = await drive_service.ensure_drive_setup(google_id)
+    return JSONResponse(result)
+
+
+@router.post("/drive/projects")
+async def create_drive_project(
+    folder_id: Optional[str] = Form(None),
+    files: List[UploadFile] = File(...),
+    google_id: Optional[str] = Query(None, description="Drive 작업에 사용할 Google 사용자 식별자 (sub)"),
+) -> Dict[str, Any]:
+    if not files:
+        raise HTTPException(status_code=422, detail="최소 한 개의 파일을 업로드해주세요.")
+
+    invalid_files: List[str] = []
+    for upload in files:
+        filename = upload.filename or "업로드된 파일"
+        if not filename.lower().endswith(".pdf"):
+            invalid_files.append(filename)
+
+    if invalid_files:
+        detail = ", ".join(invalid_files)
+        raise HTTPException(status_code=422, detail=f"PDF 파일만 업로드할 수 있습니다: {detail}")
+
+    return await drive_service.create_project(
+        folder_id=folder_id,
+        files=files,
+        google_id=google_id,
+    )
