+++ frontend/src/App.css
@@ -321,12 +321,44 @@
   display: flex;
   flex-direction: column;
   gap: 1.5rem;
+  position: relative;
 }
 
 .modal__body {
   display: flex;
   flex-direction: column;
   gap: 1.5rem;
+}
+
+.modal__loading-overlay {
+  position: absolute;
+  inset: 0;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 1rem;
+  border-radius: 18px;
+  background: rgba(248, 250, 252, 0.82);
+  backdrop-filter: blur(4px);
+  box-shadow: inset 0 0 0 1px rgba(148, 163, 184, 0.2);
+  z-index: 5;
+}
+
+.modal__loading-spinner {
+  width: 2.5rem;
+  height: 2.5rem;
+  border-radius: 999px;
+  border: 3px solid rgba(148, 163, 184, 0.35);
+  border-top-color: #2563eb;
+  animation: drive-spin 0.8s linear infinite;
+}
+
+.modal__loading-text {
+  margin: 0;
+  font-size: 1rem;
+  color: #1f2937;
+  text-align: center;
 }
 
 .modal__helper-text {
